Reuse BNACore for lambdas
Moves the common content back to the BNACore and use it in the Lambda
functions.
