+++ bnacore/src/lib.rs
@@ -1,5 +1,6 @@
 //! This crate defines the structures and functions which are shared between
 //! the PFB projects.
+pub mod aws;
 pub mod bundle;
 pub mod combine;
 pub mod neon;
