--- lambdas/src/neon/mod.rs
@@ -1,3 +1,0 @@
-pub mod model;
-
-pub const NEON_PROJECTS_URL: &str = "https://console.neon.tech/api/v2/projects";
